Move makeline and makeshape from util to private methods of bezier
