+++ lib/bezier.js
@@ -743,8 +743,8 @@
         fe = fcurves[len - 1].points[fcurves[len - 1].points.length - 1],
         bs = bcurves[len - 1].points[bcurves[len - 1].points.length - 1],
         be = bcurves[0].points[0],
-        ls = utils.makeline(bs, fs),
-        le = utils.makeline(fe, be),
+        ls = this._makeline(bs, fs),
+        le = this._makeline(fe, be),
         segments = [ls]
           .concat(fcurves)
           .concat([le])
@@ -758,7 +758,7 @@
       var outline = this.outline(d1, d2).curves;
       var shapes = [];
       for (var i = 1, len = outline.length; i < len / 2; i++) {
-        var shape = utils.makeshape(
+        var shape = this._makeshape(
           outline[i],
           outline[len - i],
           curveIntersectionThreshold
@@ -841,6 +841,48 @@
       errorThreshold = errorThreshold || 0.5;
       var circles = [];
       return this._iterate(errorThreshold, circles);
+    },
+    _makeline: function(p1, p2) {
+      var x1 = p1.x,
+        y1 = p1.y,
+        x2 = p2.x,
+        y2 = p2.y,
+        dx = (x2 - x1) / 3,
+        dy = (y2 - y1) / 3;
+      return new Bezier(
+        x1,
+        y1,
+        x1 + dx,
+        y1 + dy,
+        x1 + 2 * dx,
+        y1 + 2 * dy,
+        x2,
+        y2
+      );
+    },
+    _makeshape: function(forward, back, curveIntersectionThreshold) {
+      var bpl = back.points.length;
+      var fpl = forward.points.length;
+      var start = this._makeline(back.points[bpl - 1], forward.points[0]);
+      var end = this._makeline(forward.points[fpl - 1], back.points[0]);
+      var shape = {
+        startcap: start,
+        forward: forward,
+        back: back,
+        endcap: end,
+        bbox: utils.findbbox([start, forward, back, end])
+      };
+      var self = utils;
+      shape.intersections = function(s2) {
+        return self.shapeintersections(
+          shape,
+          shape.bbox,
+          s2,
+          s2.bbox,
+          curveIntersectionThreshold
+        );
+      };
+      return shape;
     },
     _error: function(pc, np1, s, e) {
       var q = (e - s) / 4,
