+++ lib/utils.js
@@ -344,26 +344,6 @@
       return utils.lli4(v1, v1.c, v2, v2.c);
     },
 
-    makeline: function(p1, p2) {
-      var Bezier = require("./bezier");
-      var x1 = p1.x,
-        y1 = p1.y,
-        x2 = p2.x,
-        y2 = p2.y,
-        dx = (x2 - x1) / 3,
-        dy = (y2 - y1) / 3;
-      return new Bezier(
-        x1,
-        y1,
-        x1 + dx,
-        y1 + dy,
-        x1 + 2 * dx,
-        y1 + 2 * dy,
-        x2,
-        y2
-      );
-    },
-
     findbbox: function(sections) {
       var mx = nMax,
         my = nMax,
@@ -408,31 +388,6 @@
         });
       });
       return intersections;
-    },
-
-    makeshape: function(forward, back, curveIntersectionThreshold) {
-      var bpl = back.points.length;
-      var fpl = forward.points.length;
-      var start = utils.makeline(back.points[bpl - 1], forward.points[0]);
-      var end = utils.makeline(forward.points[fpl - 1], back.points[0]);
-      var shape = {
-        startcap: start,
-        forward: forward,
-        back: back,
-        endcap: end,
-        bbox: utils.findbbox([start, forward, back, end])
-      };
-      var self = utils;
-      shape.intersections = function(s2) {
-        return self.shapeintersections(
-          shape,
-          shape.bbox,
-          s2,
-          s2.bbox,
-          curveIntersectionThreshold
-        );
-      };
-      return shape;
     },
 
     getminmax: function(curve, d, list) {
